removed search bar functionality from all the other fragments for Borrower and added only to browser
