+++ app/src/test/java/com/example/glassesgang/BrowseFragmentTest.java
@@ -1,0 +1,4 @@
+package com.example.glassesgang;
+
+public class BrowseFragmentTest {
+}
